+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to this project will be documented in this file. The format 
 - Measurements can now be rearranged by drag and drop on the measurement list
   and detail pages.
 - The date column now comes first in progress tables and CSV exports.
+- Export of progress graph (png) and raw table data (csv) are now triggered from
+  its respective widgets in the measurement details page.
 
 ### Fixed